+++ contracts/v0.2/src/interfaces/IFeedProxy.sol
@@ -6,15 +6,15 @@ import {IFeed} from "./IFeed.sol";
 interface IFeedProxy is IFeed {
     function getProposedFeed() external view returns (address);
 
-    function proposedGetRoundData(uint80 roundId)
+    function proposedGetRoundData(uint64 roundId)
         external
         view
-        returns (uint80 id, int256 answer, uint256 updatedAt);
+        returns (uint64 id, int256 answer, uint256 updatedAt);
 
     function proposedLatestRoundData()
         external
         view
-        returns (uint80 id, int256 answer, uint256 updatedAt);
+        returns (uint64 id, int256 answer, uint256 updatedAt);
 
     function getFeed() external view returns (address);
